+++ projects/ui/src/lib/components/link/index.ts
@@ -1,0 +1,3 @@
+export * from './kui-link.directive';
+export * from './kui-link-tone.type';
+export * from './kui-link-underline.type';